pyproject.toml
[tool.poetry]
name = "retail-network"
version = "0.1.0"
description = ""
authors = ["Uliana Semyonova <i@ulianasemyonova.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.0"
django-filter = "^23.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
